code cleanup & lighthouse optimization * Optimize to 100 score on lighthouse across performance and accessibility * Breakout message components into individual files
