+++ src/app/chat/_components/message-board/index.ts
@@ -1,0 +1,1 @@
+export { MessageBoard } from "./message-board";
